move TextInput to internal/form
